DataType Conversions and Code Cleanup.

+++ sku.0/sys.server/compiled/game/script/space/quest_logic/destroy_duty.java
@@ -5,7 +5,7 @@ import script.library.*;
 
 public class destroy_duty extends script.base_script
 {
-    private static final double dutyTokenBonus = Double.parseDouble(getConfigSetting("GameServer", "dutyTokenBonus"));
+    private static final float dutyTokenBonus = Float.parseFloat(getConfigSetting("GameServer", "dutyTokenBonus"));
     public static final string_id SID_TARGET_LOCATED = new string_id("space/quest", "destroy_duty_target_located");
     public static final string_id SID_TARGET_DETECTED = new string_id("space/quest", "destroy_duty_target_detected");
     public static final string_id SID_BOSS_DETECTED = new string_id("space/quest", "destroy_duty_boss_detected");
